+++ core/src/test/java/de/bitbrain/braingdx/tmx/MockObjectLayerBuilder.java
@@ -12,7 +12,7 @@ public class MockObjectLayerBuilder {
 
    private MapObjects objects = new MapObjects();
 
-   MockObjectLayerBuilder addObject(float x, float y, float size, Object type, boolean collision) {
+   MockObjectLayerBuilder addObject(float x, float y, float size, Object type, boolean collision, String producer, boolean sticky) {
       MapObject object = new MapObject();
       MapProperties properties = object.getProperties();
       properties.put(Constants.X, x);
@@ -21,22 +21,31 @@ public class MockObjectLayerBuilder {
       properties.put(Constants.HEIGHT, size);
       properties.put(Constants.TYPE, type);
       properties.put(Constants.COLLISION, collision);
+      if (producer != null) {
+         properties.put(Constants.PRODUCER, producer);
+      }
+      properties.put(Constants.STICKY, sticky);
       objects.add(object);
       return this;
    }
 
    MockObjectLayerBuilder addObject(float x, float y, Object type, boolean collision) {
-      addObject(x, y, 1f, type, collision);
+      addObject(x, y, 1f, type, collision, null, false);
       return this;
    }
 
    MockObjectLayerBuilder addObject(float x, float y, float size, Object type) {
-      addObject(x, y, size, type, true);
+      addObject(x, y, size, type, true, null, false);
       return this;
    }
 
    MockObjectLayerBuilder addObject(float x, float y, Object type) {
-      addObject(x, y, 1f, type, true);
+      addObject(x, y, 1f, type, true, null, false);
+      return this;
+   }
+
+   MockObjectLayerBuilder addObject(float x, float y, Object type, String producer, boolean sticky) {
+      addObject(x, y, 1f, type, true, producer, sticky);
       return this;
    }